Moved get_task_expert to mttl.models.library.expert_library
